+++ components/Pages/CarePackages/BrokerageBorderCost.js
@@ -8,7 +8,7 @@ const BrokerageBorderCost = ({ totalCostHeader, totalCost, className = '' }) => 
   return (
     <Container className={`brokerage__border-cost ${className}`} display='flex'>
       <p>{totalCostHeader}</p>
-      <p className='text-lbh-f01 font-weight-bold'>{minusSign}{currency.euro}{totalCost ? Math.abs(totalCost).toFixed(2) : 0}</p>
+      <p className='lbh-color-dark-red font-weight-bold'>{minusSign}{currency.euro}{totalCost ? Math.abs(totalCost).toFixed(2) : 0}</p>
     </Container>
   );
 };
