+++ components/Pages/CarePackages/ReviewPackageDetails/CareChargesSummary.js
@@ -1,0 +1,35 @@
+import React from 'react';
+import { HorizontalSeparator } from '../../../HackneyDS';
+
+const careChargesClaimCollector = {
+  2: 'Hackney Council (gross)',
+  1: 'Supplier (net)',
+};
+
+export const CareChargesSummary = ({ containerId, claimCollector, subTypeName, collectingReasonLabel }) => {
+  if (!(containerId === 'care-charges' && careChargeClaimCollector)) return null;
+
+  const careChargeClaimCollector = careChargesClaimCollector[claimCollector];
+
+  return (
+    <>
+      <p>
+        <span className="font-weight-bold">{subTypeName} (pre-assessement)</span>
+      </p>
+      <HorizontalSeparator height={8} />
+      {careChargeClaimCollector && (
+        <p>
+          <span className="font-weight-bold">Collected by: </span>
+          {careChargeClaimCollector}
+        </p>
+      )}
+      <HorizontalSeparator height={8} />
+      {collectingReasonLabel && (
+        <>
+          <p className="font-weight-bold">Why is Hackney collecting these care charges: </p>
+          <p className="mb-3">{collectingReasonLabel}</p>
+        </>
+      )}
+    </>
+  )
+}
